+++ package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "@socket.io/admin-ui": "^0.5.1",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "socket.io": "^4.7.2"
   },
